➕Added Home Screen Left Rectangle with its details.
Implemented the left rectangle with it displaying a description and some clickable icons for other social media. The social media icons each are an instance of the ClickableCircleImage widget which will open a new tab with a corresponding link when clicked. Also, updated the constants file.
NOTE: The image assets for icons (and the top menu bar) must later be modified for better quality.