➕Added use-cases.
Every UseCase will be a child of the UseCase class by indicating its data Type and related Parameters. NoParams is to be used whenever a UseCase doesn't take any arguments in its call Method (See GetSkills usecase as an example).
Finally added a GetSkills usecase which is responsible for retrieving a set of skills from the server.